requirements.txt
asgiref==3.5.2
Django==4.1
sqlparse==0.4.3
whitenoise==6.2.0
django-on-heroku==1.1.2
gunicorn==20.1.0
djangorestframework==3.14.0
requests==2.28.1
django-crispy-forms==1.14.0